Fixed bug with printing
Use UserControls instead of DataTemplate. Background:
With DataTemplates, when the last alarm was ack'ed, it somehow was still
cached by the application. When a new alarm came in and was printed, the
PrintVisual()-function would not print the new alarm, but the previous
one. This commit is fixing this behavior. If a better solution is known
-> go on :)