OET does not throw on unsupported expression anymore
But what it will do however, is - if desired - manually throw an
exception that gives more details. The reason is that currently,
retrieving the value from an indexer is not supported (if that will ever
be supported). Thus what we now do, is yielding an exception that is
more helpful.