More robust way of printing copies
Some printers don't allow printing more than a specified number of
copies. For example, the Microsoft XPS Writer has a MaximumCopies-Value
of "1", meaning a user that configures a copy count of "5" in the
editor, will still have only 1 copy being printed.

This change allows for printing multiple copies, by running the printing
task multiple times. It is only done this way when the printer does not
allow as much copies. In other cases, where the printer does natively
support printing x copies, it is done this way then.