New algorithm for printing
Using FixedDocument and FixedPage. There seems to be trouble printing
with WPF, but this method seems to be reasonable. We use the
FixedXYZ-classes to print with a DocumentPaginator. A VisualBrush is
used to create a copy of the control for printing. The VisualBrush is
stretched using "Uniform" method. We don't use "Fill" because that would
stretch the picture unnaturally. "Uniform" is a good compromise (it also
saves some printing ink here by the way ;)).

Got inspiration from
"http://social.msdn.microsoft.com/Forums/en-US/wpf/thread/3167e940-03fa-4ce2-a002-6b28c60328bf/".