Basic error report management integrated
The error report management features a set of methods which can be used
to record exceptions on a general purpose-level. This means any
exception that occurs anywhere can be tracked. The intent of this
management however is, that any unhandled exception of the app domain is
tracked. The reason for this is to get a clear distinction between
runtime errors which are potentionally "ok" to those that were not
caught correctly and escalate all the way up to the AppDomain and -
possibly - terminate the process, which is something we don't want. But
we also don't want to simply swallow each exception.

The ERM targets all issues and also serves as a future extension point
to potentionally allow to send the error reports to the OFS team.