OperationPrinter working
Uses a customizable HTML-page to layout the contents. No special magic
here other than the backstage-usage of ObjectFormatter to parse the
HTML-file and replace the macros. This makes the page fully customizable
without deeper programming knowledge, except maybe basic HTML.

The route-plan-image is realized as a special HTML-feature called
"embedded img", which embeds the Base64-encoded binary image file
contents. This allows us to circumvent some shortcomings with the
WebBrowser-control, which does not reliably (at least not on my two
testing PCs) parse/include a <img src="" /> that links to a file in the
same directory as the HTML page. Using the embedded Base64-contents is
not really slower but works bulletproof - at least as far as I could
test.
In case of an error in retrieving the route image, or GoogleMaps not
finding a result, a placeholder image is displayed - neat.