Fixed apparent merge-error in Operation Store
Don't use the Alarm-time when persisting an Operation. This is dangerous because it can mess up sorting (especially when developing), and is technically wrong too. The "Timestamp"-column is currently interpreted as the timestamp of the alarm-income, not the actual alarm time. This will be cleaned up when the new database structure is finished!