[Config UI] Tweaked usability with service installation
Only allow controlling the service when the running ConfEd is configured to run on "localhost". Reason: it doesn't make sense to allow a connected client to install and control a service. This should be allowed on the server only.
Also removed the context menu and added a drop down menu with a visual cue to make it more obvious.