Added configuration to ease debugging the website
A new configuration setting was introduced, called "Debug.AlarmWorkflowLocation", which, when set, points to the local AlarmWorkflow\Build-directory. This is only applied when:
1. The website was built in DEBUG configuration
2. The website is run with a debugger attached

If the path does not exist or is blank, the application path (where the website is installed at) will be used.

If any is not the case, then the application path will be used, just like it has been so far.