Added generic exception handling to website
Any uncaught exceptions are automatically logged and a redirection to the error-page is done.

This happens in all cases. "mode" is set to "On" (and not "removeOnly") because the error page displayed when debugging is not too helpful and also it is not guaranteed that people host a dedicated server and display the website on a different machine.