Moved service management into the Service application
Removed it from the configuration. It is less confusing that way - if a client is connected to a server, the "service status" is confusing as it is not transmitted to the client (and shouldn't be in the first place since it's only important for the Server).

The AlarmWorkflow.Backend.Service.exe has to be started *as an administrator* to be able to use it.